# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API's",
        "description": "How to login:\n *   Login using the `/authentication/login` endpoint with your email and password.\n *   Copy the token from the response, specifically from `data.token.value`.\n *   Click the Authorize button in the top right corner.\n *   Enter the token in the format: `Bearer <access_token>`.\n *   Click the Authorize button.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1/",
            "description": "LIVE server"
        }
    ],
    "paths": {
        "/authentication/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login to systems and return the token and inits data.",
                "description": "Do Login",
                "operationId": "a0b5083c5b1dfa24e85eadffbc31c920",
                "parameters": [
                    {
                        "name": "body",
                        "in": "query",
                        "description": "Login Username & Password",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "example": "admin@mail.test"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "123456"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/authentication/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user.",
                "description": "Do Logout",
                "operationId": "c3f856105a08dfdd9274d4c4ad74339d",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CreateRole": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "New Role Name"
                    },
                    "permission_ids": {
                        "type": "integer",
                        "example": [
                            1,
                            2
                        ]
                    }
                },
                "type": "object"
            },
            "UpdateRole": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "New Role Update"
                    },
                    "permission_ids": {
                        "type": "integer",
                        "example": [
                            1,
                            2
                        ]
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        }
    ]
}